+++ src/zig_clang.h
@@ -8,7 +8,7 @@
 #ifndef ZIG_ZIG_CLANG_H
 #define ZIG_ZIG_CLANG_H
 
-#include "userland.h"
+#include "stage2.h"
 #include <inttypes.h>
 #include <stdbool.h>
